+++ API/Repositories/ICourseRepository.cs
@@ -9,6 +9,6 @@ namespace API.Repositories
     public interface ICourseRepository
     {
         Task<IEnumerable<Course>> GetAll();
-        Task<Course> GetCourseById(long id);
+        Task<Course> GetCourseById(int id);
     }
 }
